Revert "Revert "Added documentation to new InformationGainFormula class.""
This reverts commit 77dba7f9ec3a18f3343b2e39fccf8140f3731abd [formerly 605727f713cc13a9b40d7dd11b1c6373c404f8f7].


Former-commit-id: 47bb3f9078dfd85f9168adcfc7fee9b9f33367a1

+++ palladian-retrieval/src/main/java/ws/palladian/retrieval/search/Searcher.java
@@ -36,7 +36,7 @@ public interface Searcher<R extends SearchResult> {
      * @return
      * @throws SearcherException In case the search fails.
      */
-    long getTotalResultCount(String query) throws SearcherException;
+    int getTotalResultCount(String query) throws SearcherException;
 
     /**
      * <p>
